test: refactor test262
