+++ package.json
@@ -75,7 +75,7 @@
     "postversion": "git push && git push --tags && npm publish",
     "prepare": "husky",
     "generate-unicode": "node scripts/generate-unicode.mjs",
-    "generate-test262-whitelist": "node scripts/generate-test262-whitelist.mjs"
+    "generate-test262-whitelist": "npm run bundle && node scripts/generate-test262-whitelist.mjs"
   },
   "devDependencies": {
     "@eslint/js": "^9.24.0",
@@ -100,7 +100,6 @@
     "standard-changelog": "^6.0.0",
     "test262": "github:tc39/test262#afc3d1bce4ae643e74296fb21d8665ec43a1f2bd",
     "test262-parser-runner": "^0.5.0",
-    "ts-node": "^10.9.2",
     "typescript": "^5.8.2",
     "typescript-eslint": "^8.27.0",
     "vitest": "^3.1.1"
