+++ scripts/run-test262-cli.mjs
@@ -1,0 +1,4 @@
+import runTest from './run-test262.mjs';
+import * as meriyah from '../dist/meriyah.mjs';
+
+await runTest(meriyah);
