--- scripts/run-test262-cli.ts
@@ -1,3 +1,0 @@
-import runTest from './run-test262';
-
-runTest();
